Move preview code to its own file and remove :cat2: :cat2:
