+++ src/extension.ts
@@ -3,8 +3,8 @@ import { validate } from './server/messages';
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import * as os from 'os';
 import * as nisabaLogger from './logger';
+import { PreviewPanel } from './preview';
 
 // Logging output channel
 const nisabaOutputChannel = vscode.window.createOutputChannel("Nisaba");
@@ -42,7 +42,7 @@ export function activate(context: vscode.ExtensionContext) {
 
     context.subscriptions.push(
         vscode.commands.registerCommand('ucl-rsdg.arabicPreview', () => {
-            CatCodingPanel.createOrShow(context.extensionUri);
+            PreviewPanel.createOrShow(context.extensionUri);
         })
     );
 }
@@ -50,181 +50,4 @@ export function activate(context: vscode.ExtensionContext) {
 // this method is called when your extension is deactivated
 export function deactivate() {
     nisabaLogger.stopLogging();
-}
-
-
-class CatCodingPanel {
-    /**
-     * Track the currently panel. Only allow a single panel to exist at a time.
-     */
-    public static currentPanel: CatCodingPanel | undefined;
-
-    public static readonly viewType = 'catCoding';
-
-    private readonly _panel: vscode.WebviewPanel;
-    private readonly _extensionUri: vscode.Uri;
-    private _disposables: vscode.Disposable[] = [];
-    private _document: vscode.TextDocument;
-
-    public static createOrShow(extensionUri: vscode.Uri) {
-        const column = vscode.window.activeTextEditor
-            ? vscode.window.activeTextEditor.viewColumn
-            : undefined;
-
-        // If we already have a panel, show it.
-        if (CatCodingPanel.currentPanel) {
-            CatCodingPanel.currentPanel._panel.reveal(column);
-            return;
-        }
-
-        // Otherwise, create a new panel.
-        const panel = vscode.window.createWebviewPanel(
-            CatCodingPanel.viewType,
-            'Cat Coding',
-            vscode.ViewColumn.Two,
-            {
-                // Enable javascript in the webview
-                enableScripts: true,
-
-                // And restrict the webview to only loading content from our extension's `media` directory.
-                localResourceRoots: [vscode.Uri.joinPath(extensionUri, 'media')]
-            }
-        );
-
-        CatCodingPanel.currentPanel = new CatCodingPanel(panel, extensionUri);
-    }
-
-    public static revive(panel: vscode.WebviewPanel, extensionUri: vscode.Uri) {
-        CatCodingPanel.currentPanel = new CatCodingPanel(panel, extensionUri);
-    }
-
-    private constructor(panel: vscode.WebviewPanel, extensionUri: vscode.Uri) {
-        this._panel = panel;
-        this._extensionUri = extensionUri;
-
-        // Set the webview's initial html content
-        this._document = vscode.window.activeTextEditor.document;
-        this._update();
-
-        // Update the webview based on text changes
-        const changeDocumentSubscription = vscode.workspace.onDidChangeTextDocument(e => {
-            if (e.document.uri.toString() === this._document.uri.toString()) {
-                this._update();
-            }
-        });
-
-        // Make sure we get rid of the listener when our editor is closed.
-        // TODO Should we be using the extra arguments like in the subscription below?
-        // Or push this into _disposables instead?
-        // Do the multiple calls to onDidDispose stack or overwrite the listeners?
-        this._panel.onDidDispose(() => {
-            changeDocumentSubscription.dispose();
-        });
-
-        const switchDocumentSubscription = vscode.window.onDidChangeActiveTextEditor(e => {
-            if (e.document == undefined) {
-                this._document = e.document;
-                // TODO Delete contents?
-            } else {
-                this._document = e.document;
-                this._update();
-            }
-        });
-        this._panel.onDidDispose(() => switchDocumentSubscription.dispose());
-
-        // Listen for when the panel is disposed
-        // This happens when the user closes the panel or when the panel is closed programatically
-        this._panel.onDidDispose(() => this.dispose(), null, this._disposables);
-
-        // Update the content based on view changes
-        this._panel.onDidChangeViewState(
-            () => {
-                if (this._panel.visible) {
-                    this._update();
-                }
-            },
-            null,
-            this._disposables
-        );
-
-        // Handle messages from the webview
-        this._panel.webview.onDidReceiveMessage(
-            message => {
-                switch (message.command) {
-                    case 'alert':
-                        vscode.window.showErrorMessage(message.text);
-                        return;
-                }
-            },
-            null,
-            this._disposables
-        );
-    }
-
-    public doRefactor() {
-        // Send a message to the webview webview.
-        // You can send any JSON serializable data.
-        this._panel.webview.postMessage({ command: 'refactor' });
-    }
-
-    public dispose() {
-        CatCodingPanel.currentPanel = undefined;
-
-        // Clean up our resources
-        this._panel.dispose();
-
-        while (this._disposables.length) {
-            const x = this._disposables.pop();
-            if (x) {
-                x.dispose();
-            }
-        }
-    }
-
-    private _update() {
-        const webview = this._panel.webview;
-        this._updateForCat(webview);
-    }
-
-    private _updateForCat(webview: vscode.Webview) {
-        this._panel.title = "Oracc Preview";
-        this._panel.webview.html = this._getHtmlForWebview(webview, this._document);
-    }
-
-    private _getHtmlForWebview(webview: vscode.Webview, document: vscode.TextDocument) {
-        const lines = document.getText().split(os.EOL);
-        let arabic = false;
-        for (let i = 0; i < lines.length; i++) {
-            // `dir` tag: by default we assume left-to-right
-            let dir = "dir=\"ltr\"";
-            if (lines[i].match(/^@translation .* ar/)) {
-                // We start an Arabic translation
-                arabic = true;
-            }
-            if (arabic && lines[i].match(/^\d+\..*/)) {
-                // We're inside Arabic translation and this is a text line: set
-                // `dir` to right-to-left.
-                dir = "dir=\"rtl\"";
-            }
-            lines[i] = lines[i].replace(/^(.*)$/, `<p ${dir}>$1</p>`);
-        }
-        return `<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-
-<!--
-Use a content security policy to only allow loading images from https or from our extension directory,
-and only allow scripts that have a specific nonce.
--->
-<meta http-equiv="Content-Security-Policy" content="default-src 'none'; img-src ${webview.cspSource} https:;">
-
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>Cat Coding</title>
-</head>
-<body>
-${lines.join(os.EOL)}
-</body>
-</html>`;
-    }
 }
